requirements.txt
pillow
numpy
pillow-avif-plugin
ttkbootstrap
imagecodecs
pyinstaller